Natural language requirements extractions and judgment (#18)
Adds natural language requirement lists to verisafe.

Unless requirements are skipped (via the --skip-reqs flag), the system document and rule file are analyzed by an agent (called the extractor), which has been prompted to ask a large number of questions of the user to clarify certain design decisions. This agent then outputs a list of requirements, which are included as part of the initial prompt to the workflow. The extractor also has access to a memory tool, which it is instructed to update before it produces it's list of requirements.

In a resume workflow from some prior workflow T, the previously analyzed spec file is also fed into the extractor, and its memory is initialized with the memories from T.

In addition, we explicitly track which validations which are required for the workflow to complete. If requirements are enabled, checking the natural language requirements are added to the set of required validations.

After code generation, the synthesis agent must call the "requirements evaluator" tool, or the "judge". This judge is provided access to the VFS and the natural language requirements and is instructed to evaluate each requirement against the VFS. This judge has access to the memories generated by the extractor. It will return its evaluation to the synthesis agent, which is instructed to address any feedback which is not SATISFIED or LIKELY. If all requirements which have not been relaxed (see below) are SATISFIED or LIKELY, then the validation state is updated to record that the current implementation passes the requirements.

However, because these are natural language requirements being interpreted by an LLM, we provide an escape hatch, the requirements relaxation tool. This allows the synthesis agent to request that some of the requirements be relaxed, which allows completion of the code synthesis task even if some of the requirements are not satisfied or likely.

This PR includes a lot of refactoring to cut down on boilerplate. The human in the loop and the requirements relaxation request tools have the same shape; that shared infrastructure has been collapsed into a "human_interaction_tool" factory. The result tool of the synthesis agent has been updated to use the new validation functionality of the result tool letting us delete some code. In addition, so we don't go insane while debugging, I implemented a third agent that pretends to be a human answering the extractors question by analyzing an existing implementation. This is called the requirements oracle, and is expected to only be useful in our debugging/dev cycles.