Debug console (#20)
When hitting Ctrl+C this will set a flag indicating a request for the debug console. After the next tools node, the user is dropped
into a debug console where they can review the messages or VFS. They can also provide explicit direction as a human message.

The user can also write DEBUG at any of the human interaction points and be dropped into a shell, without the option of providing explicit feedback.